+++ apps/hub/src/components/NewModelButton.tsx
@@ -1,5 +1,3 @@
-"use client";
-import { useSelectedLayoutSegment } from "next/navigation";
 import { FC } from "react";
 
 import { PlusIcon } from "@quri/ui";
@@ -7,17 +5,12 @@ import { PlusIcon } from "@quri/ui";
 import { LinkButton } from "@/components/ui/LinkButton";
 import { newModelRoute } from "@/lib/routes";
 
-export const NewModelButton: FC<{ group: string }> = ({ group }) => {
-  const segment = useSelectedLayoutSegment();
-
+// used on the group page, user page and frontpage
+export const NewModelButton: FC<{ group?: string }> = ({ group }) => {
   const link = newModelRoute({ group });
 
-  if (segment === "members" || segment === "invite-link") {
-    return null;
-  }
-
   return (
-    <LinkButton href={link}>
+    <LinkButton href={link} theme="primary">
       <div className="flex items-center gap-1">
         <PlusIcon size={16} />
         New Model
